# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"DidDocument": {
				"$ref": "#/components/schemas/Record_string.any_"
			},
...
			"KeyAlgorithm": {
				"enum": [
					"a128gcm",
					"a256gcm",
					"a128cbchs256",
					"a256cbchs512",
					"a128kw",
					"a256kw",
					"bls12381g1",
					"bls12381g2",
					"c20p",
					"xc20p",
					"ed25519",
					"x25519",
					"k256",
					"p256",
					"p384"
				],
				"type": "string"
			},
			"DidCreate": {
				"properties": {
					"keyType": {
						"$ref": "#/components/schemas/KeyAlgorithm"
					},
					"seed": {
						"type": "string"
					},
					"domain": {
						"type": "string"
					},
					"method": {
						"type": "string"
					},
					"network": {
						"type": "string"
					},
					"did": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"endorserDid": {
						"type": "string"
					},
					"didDocument": {
						"$ref": "#/components/schemas/DidDocument"
					},
					"privatekey": {
						"type": "string"
					},
					"endpoint": {
						"type": "string"
					},
					"isDefault": {
						"type": "boolean"
					}
				},
				"required": [
					"method"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidRecord": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
...
			"AgentInfo": {
				"properties": {
					"label": {
						"type": "string"
					},
					"endpoints": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isInitialized": {
						"type": "boolean"
					},
					"publicDid": {}
				},
				"required": [
					"label",
					"endpoints",
					"isInitialized",
					"publicDid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AgentToken": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyDataOptions": {
				"properties": {
					"publicKeyBase58": {
						"type": "string",
						"description": "Optional; ignored for verification — the DID-resolved key is always used"
					},
					"method": {
						"type": "string",
						"description": "Optional DID method hint (informational)"
					},
					"keyType": {
						"type": "string",
						"description": "Key type hint (informational, e.g. \"ed25519\")"
					},
					"signature": {
						"type": "string",
						"description": "Signature to verify, base64 encoded"
					},
					"data": {
						"type": "string",
						"description": "Signed data, base64 encoded"
					},
					"did": {
						"type": "string",
						"description": "DID whose authentication key verifies the signature. The verification key is\nresolved from the DID Document, never taken from the request body."
					}
				},
				"required": [
					"signature",
					"data",
					"did"
				],
				"type": "object"
			},
...
			"W3cVerifyCredentialResult": {
				"$ref": "#/components/schemas/W3cVerifyResult_W3cCredentialValidations_"
			},
...
			"SingleOrArray_any-or-DataIntegrityProofOptions_": {
				"anyOf": [
					{
						"anyOf": [
							{},
							{
								"$ref": "#/components/schemas/DataIntegrityProofOptions"
							}
						]
					},
					{
						"items": {
							"anyOf": [
								{},
								{
									"$ref": "#/components/schemas/DataIntegrityProofOptions"
								}
							]
						},
						"type": "array"
					}
				]
			},
...
			"ProofPurpose": {},
			"SafeW3cJsonLdVerifyCredentialOptions": {
				"properties": {
					"credential": {
						"$ref": "#/components/schemas/W3cJsonLdVerifiableCredential"
					},
					"verifyCredentialStatus": {
						"type": "boolean",
						"description": "Whether to verify the credentialStatus, if present."
					},
					"proofPurpose": {
						"$ref": "#/components/schemas/ProofPurpose"
					},
					"proof": {
						"$ref": "#/components/schemas/SingleOrArray_any-or-DataIntegrityProofOptions_"
					}
				},
				"required": [
					"credential",
					"proof"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SelfAttestedW3cCredentialResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"credentialInstances": {
						"items": {
							"properties": {
								"credential": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/JsonObject"
										},
										{
											"type": "string"
										}
									]
								}
							},
							"required": [
								"credential"
							],
							"type": "object"
						},
						"type": "array"
					},
					"credential": {
						"$ref": "#/components/schemas/JsonObject"
					}
				},
				"required": [
					"id",
					"createdAt",
					"credentialInstances",
					"credential"
				],
				"type": "object",
				"additionalProperties": {}
			},
...
			"jsonLdCredentialOptions": {
				"properties": {
					"@context": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"$ref": "#/components/schemas/JsonObject"
								}
							]
						},
						"type": "array"
					},
					"type": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"credentialSubject": {
						"$ref": "#/components/schemas/SingleOrArray_JsonObject_"
					},
					"proofType": {
						"type": "string"
					},
					"expirationDate": {
						"type": "string"
					}
				},
				"required": [
					"@context",
					"type",
					"credentialSubject",
					"proofType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "credo-controller",
		"version": "2.1.3",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/dids/write": {
			"post": {
				"operationId": "WriteDid",
				"responses": {
					"200": {
						"description": "DidResolutionResult",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"did": "did:indy:bcovrin:testnet:LRCUFcizUL74AGgLqdJHK7",
											"didDocument": {
												"@context": [
													"https://w3id.org/did/v1",
													"https://w3id.org/security/suites/ed25519-2018/v1"
												],
												"id": "did:indy:bcovrin:testnet:LRCUFcizUL74AGgLqdJHK7",
												"verificationMethod": [
													{
														"id": "did:indy:bcovrin:testnet:LRCUFcizUL74AGgLqdJHK7#verkey",
														"type": "Ed25519VerificationKey2018",
														"controller": "did:indy:bcovrin:testnet:LRCUFcizUL74AGgLqdJHK7",
														"publicKeyBase58": "BapLDK4dEY88vWcQgNbpAPVVP4r3CHs4MvShmmhqkxXM"
													}
												],
												"authentication": [
													"did:indy:bcovrin:testnet:LRCUFcizUL74AGgLqdJHK7#verkey"
												]
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Did nym registration",
				"tags": [
					"Dids"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCreate"
							}
						}
					}
				}
			}
		},
		"/dids": {
			"get": {
				"operationId": "GetDids",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DidRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Dids"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "isDefault",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
		"/agent": {
			"get": {
				"operationId": "GetAgentInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AgentInfo"
								}
							}
						}
					}
				},
				"description": "Retrieve basic agent information",
				"tags": [
					"Agent"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated",
							"Basewallet"
						]
					}
				],
				"parameters": []
			}
		},
		"/agent/token": {
			"post": {
				"operationId": "GetAgentToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AgentToken"
								}
							}
						}
					}
				},
				"description": "Retrieve agent token",
				"tags": [
					"Agent"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": []
			}
		},
		"/agent/verify": {
			"post": {
				"operationId": "Verify",
				"responses": {
					"200": {
						"description": "true if the signature is valid for the DID-resolved key, false otherwise",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Verify a DID-bound challenge-response signature.\n\nThe signing algorithm is determined by the key type in the holder's DID Document\n(e.g. EdDSA for Ed25519 did:key, ES256 for P-256) — the endpoint is not\nrestricted to Ed25519. Any algorithm supported by the resolved authentication\nkey is accepted.",
				"tags": [
					"Agent"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Verify options\ndid - DID whose authentication key is used to verify (resolved server-side)\ndata - Signed data, base64 encoded\nsignature - Signature to verify, base64 encoded",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyDataOptions",
								"description": "Verify options\ndid - DID whose authentication key is used to verify (resolved server-side)\ndata - Signed data, base64 encoded\nsignature - Signature to verify, base64 encoded"
							}
						}
					}
				}
			}
		},
		"/agent/credential/verify": {
			"post": {
				"operationId": "VerifyCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/W3cVerifyCredentialResult"
								}
							}
						}
					}
				},
				"tags": [
					"Agent"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"anyOf": [
									{
										"$ref": "#/components/schemas/SafeW3cJsonLdVerifyCredentialOptions"
									},
									{}
								]
							}
						}
					}
				}
			}
		},
		"/agent/credential/self-attested": {
			"post": {
				"operationId": "CreateW3cSelfAttestedCredential",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SelfAttestedW3cCredentialResponse"
								}
							}
						}
					}
				},
				"description": "Create and store a self-attested W3C JSON-LD credential — the agent issues a credential to\nitself using its own default DID, rather than an external issuer signing over a subject DID.\n\nLives on request.agent, not /multi-tenancy/:tenantId + withTenantAgent() — matching\nverifyCredential/verify above rather than the legacy pipeline-implementation placement. The\nauth layer already resolves the tenant agent per request for a tenant token (see\nauthentication.ts), which is why no live controller on develop calls withTenantAgent(); the\n/multi-tenancy/:tenantId placement is base-wallet-token only (MULTITENANT_BASE_AGENT), so a\ndedicated agent or a tenant's own token could never have reached it there.\n\nPorted from pipeline-implementation, adapted for the current Credo version:\n- request.agent instead of the legacy this.agent field (no longer exists on this controller).\n- credentialSubject's claims are nested under a `claims` key when building the subject —\n  W3cCredentialSubject's constructor only reads options.id/options.claims, so spreading the\n  request's claims at the top level (as the legacy code effectively did, via a plain object\n  that bypassed the class transform) would silently drop every claim from the issued\n  credential. Handles a single object or an array of subjects.\n- w3cCredentials.store({ record }) instead of the removed storeCredential({ credential });\n  W3cCredentialRecord.fromCredential(signedCred) builds the record directly from the signed\n  credential — signCredential() already returns a typed W3cVerifiableCredential instance, so\n  the legacy JsonTransformer.fromJSON re-parse step is no longer needed either.\n- the response restores a top-level `credential` field for back-compat: 0.6.2's\n  W3cCredentialRecord replaces 0.5.x's top-level `credential` with a `credentialInstances`\n  array (credential survives only as a private write-only setter), which would otherwise\n  silently change the response shape for downstream consumers (platform's apps/cloud-wallet\n  -> agent-service -> mobile wallet) that read response.credential directly.",
				"tags": [
					"Agent"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/jsonLdCredentialOptions"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}